Use "?" and ":?" parameter expansions when possible

+++ resources/stats/suffix.sh
@@ -10,9 +10,9 @@ export LC_ALL=C
 # Check if a program exists
 exists() {
 	# shellcheck disable=SC2230
-	if command -v true; then command -v -- "$1"
-	elif eval type type; then eval type -- "$1"
-	else which -- "$1"; fi >/dev/null 2>&1
+	if command -v true; then command -v -- "${1:?}"
+	elif eval type type; then eval type -- "${1:?}"
+	else which -- "${1:?}"; fi >/dev/null 2>&1
 }
 
 # Create temporary file
@@ -21,15 +21,15 @@ createTempFile() {
 	else # Since POSIX does not specify mktemp utility, use this as fallback
 		tempCounter=${tempCounter:-9999}
 		tempFile="${TMPDIR:-/tmp}/hblock.$$.$((tempCounter+=1))"
-		rm -f -- "$tempFile" && (umask 077 && touch -- "$tempFile")
-		printf -- '%s\n' "$tempFile"
+		rm -f -- "${tempFile:?}" && (umask 077 && touch -- "${tempFile:?}")
+		printf -- '%s\n' "${tempFile:?}"
 	fi
 }
 
 # Print to stdout the contents of a URL
 fetchUrl() {
-	if exists curl; then curl -fsSL -- "$1";
-	elif exists wget; then wget -qO- -- "$1";
+	if exists curl; then curl -fsSL -- "${1:?}";
+	elif exists wget; then wget -qO- -- "${1:?}";
 	else
 		logError 'Either wget or curl are required for this script'
 		exit 1
@@ -46,8 +46,8 @@ main() {
 	file="${1:?}"
 	publicSuffixList="${2:-https://publicsuffix.org/list/public_suffix_list.dat}"
 
-	if ! [ -f "$file" ] || ! [ -r "$file" ]; then
-		>&2 printf -- '%s\n' "Cannot read file '$file'"
+	if ! [ -f "${file:?}" ] || ! [ -r "${file:?}" ]; then
+		>&2 printf -- '%s\n' "Cannot read file '${file:?}'"
 		exit 1
 	fi
 
@@ -55,58 +55,58 @@ main() {
 	stats=''
 
 	# Create temporary blocklist file
-	blocklist=$(createTempFile)
-	cp -f -- "$file" "$blocklist"
-	rmtemp() { rm -f -- "$blocklist" "$blocklist".*; }
+	blocklistFile=$(createTempFile)
+	cp -f -- "${file:?}" "${blocklistFile:?}"
+	rmtemp() { rm -f -- "${blocklistFile:?}" "${blocklistFile:?}".*; }
 	trap rmtemp EXIT
 
 	# Compact blocklist content (remove lowest level domain and count ocurrences)
-	sed -e 's/^.\{1,\}[[:blank:]][^.]\{1,\}//' -- "$blocklist" \
-		| sort | uniq -c > "$blocklist.aux" \
-		&& mv -f -- "$blocklist.aux" "$blocklist"
+	sed -e 's/^.\{1,\}[[:blank:]][^.]\{1,\}//' -- "${blocklistFile:?}" \
+		| sort | uniq -c > "${blocklistFile:?}.aux" \
+		&& mv -f -- "${blocklistFile:?}.aux" "${blocklistFile:?}"
 
-	if [ "$publicSuffixList" != 'none' ]; then
+	if [ "${publicSuffixList:?}" != 'none' ]; then
 		# Download public suffix list
-		fetchUrl "$publicSuffixList" > "$blocklist.suffixes"
+		fetchUrl "${publicSuffixList:?}" > "${blocklistFile:?}.suffixes"
 
 		# Transform suffix list (punycode encode and sort by length in descending order)
-		sed -e '/^\/\//d;/^!/d;/^$/d;s/^\*\.//g' -- "$blocklist.suffixes" \
+		sed -e '/^\/\//d;/^!/d;/^$/d;s/^\*\.//g' -- "${blocklistFile:?}.suffixes" \
 			| punycodeEncode | awk '{print(length($0)":."$0)}' \
-			| sort -nr | cut -d: -f2 > "$blocklist.aux" \
-			&& mv -f -- "$blocklist.aux" "$blocklist.suffixes"
+			| sort -nr | cut -d: -f2 > "${blocklistFile:?}.aux" \
+			&& mv -f -- "${blocklistFile:?}.aux" "${blocklistFile:?}.suffixes"
 
 		# Create regex pattern for each suffix
 		sed -e 's/\./\\./g;s/$/$/g' \
-			-- "$blocklist.suffixes" > "$blocklist.aux" \
-			&& mv -f -- "$blocklist.aux" "$blocklist.suffixes"
+			-- "${blocklistFile:?}.suffixes" > "${blocklistFile:?}.aux" \
+			&& mv -f -- "${blocklistFile:?}.aux" "${blocklistFile:?}.suffixes"
 
 		# Count blocklist matches for each suffix
 		while read -r regex; do
-			if grep -- "$regex" "$blocklist" > "$blocklist.match"; then
-				count=$(awk '{s+=$1}END{print(s)}' "$blocklist.match")
-				stats=$(printf -- '%s\t%s\n%s' "$count" "$regex" "$stats")
-				(grep -v -- "$regex" "$blocklist" > "$blocklist.aux" \
-					&& mv -f -- "$blocklist.aux" "$blocklist") || true
+			if grep -- "${regex:?}" "${blocklistFile:?}" > "${blocklistFile:?}.match"; then
+				count=$(awk '{s+=$1}END{print(s)}' "${blocklistFile:?}.match")
+				stats=$(printf -- '%s\t%s\n%s' "${count:?}" "${regex:?}" "${stats?}")
+				(grep -v -- "${regex:?}" "${blocklistFile:?}" > "${blocklistFile:?}.aux" \
+					&& mv -f -- "${blocklistFile:?}.aux" "${blocklistFile:?}") || true
 			fi
-		done < "$blocklist.suffixes"
+		done < "${blocklistFile:?}.suffixes"
 
 		# Undo regex pattern
-		stats=$(printf -- '%s' "$stats" | sed 's/\\\././g;s/\$$//g')
+		stats=$(printf -- '%s' "${stats?}" | sed 's/\\\././g;s/\$$//g')
 	fi
 
 	# If blocklist is not empty use TLD as suffix
-	if [ -s "$blocklist" ]; then
-		tldStats=$(sed -e 's/^\(.\{1,\}[[:blank:]]\).*\(\.[^.]\{1,\}\)$/\1\2/g' -- "$blocklist" |
+	if [ -s "${blocklistFile:?}" ]; then
+		tldStats=$(sed -e 's/^\(.\{1,\}[[:blank:]]\).*\(\.[^.]\{1,\}\)$/\1\2/g' -- "${blocklistFile:?}" |
 			awk '{arr[$2]+=$1;}END{for (i in arr) print(arr[i]"\t"i)}'
 		)
 
-		stats=$(printf -- '%s\n%s' "$tldStats" "$stats")
+		stats=$(printf -- '%s\n%s' "${tldStats?}" "${stats?}")
 	fi
 
 	# Sort stats by the number of matches
-	stats=$(printf -- '%s' "$stats" | sort -k1,1nr -k2,2 | awk '{print NR"\t"$0}')
+	stats=$(printf -- '%s' "${stats?}" | sort -k1,1nr -k2,2 | awk '{print NR"\t"$0}')
 
-	printf -- '%s\n%s\n' "$header" "$stats"
+	printf -- '%s\n%s\n' "${header:?}" "${stats?}"
 }
 
 main "$@"
